django/requirements.txt
celery==5.3.6
crispy-bootstrap5==2024.2
Django==5.0.4
django-crispy-forms==2.1
django-environ==0.11.2
django-parler==2.3
django-rosetta==0.10.0
djangorestframework==3.15.1
flower==2.0.1
gunicorn==22.0.0
ipython==8.24.0
persiantools==4.0.0
pillow==10.3.0
psycopg2==2.9.9
redis==5.0.3
weasyprint==61.2
zeep==4.2.1
